move `deps`
